chore: Refactor web.go into 3 files (#763)
* Refactor web.go into api/schema.go

* Copy init() fn to schema.go
