+++ webv2/config.json
@@ -1,4 +1,4 @@
 {
- "GCPProjectID": "",
- "SpannerInstanceID": ""
+ "GCPProjectID": "span-cloud-testing",
+ "SpannerInstanceID": "deep-heavy-100gb"
 }
